Use more involved DOM manipulation
Instead of using regular expressions to search and replace, use a
solution that will walk the DOM and create/remove DOM nodes. The main
reason for this change is that we want to link to pivotal stories for
text that are already inside links. Regardless, this solution is more
robust and flexible.